chore: organize jest tests into tranches
