+++ test/jest/unit/format-test-results.spec.ts
@@ -1,7 +1,7 @@
-import * as jsonModule from '../src/lib/json';
+import * as jsonModule from '../../../src/lib/json';
 
-import { extractDataToSendFromResults } from '../src/cli/commands/test/formatters/format-test-results';
-import { Options } from '../src/lib/types';
+import { extractDataToSendFromResults } from '../../../src/cli/commands/test/formatters/format-test-results';
+import { Options } from '../../../src/lib/types';
 import * as fs from 'fs';
 import { deepEqual } from 'assert';
