+++ test/jest/unit/snyk-code-test.spec.ts
@@ -5,13 +5,13 @@ import { analyzeFolders, AnalysisSeverity } from '@snyk/code-client';
 jest.mock('@snyk/code-client');
 const analyzeFoldersMock = analyzeFolders as jest.Mock;
 
-import { loadJson } from './utils';
-import * as featureFlags from '../src/lib/feature-flags';
-import { config as userConfig } from '../src/lib/user-config';
-import * as analysis from '../src/lib/plugins/sast/analysis';
-import { Options, TestOptions } from '../src/lib/types';
-import * as ecosystems from '../src/lib/ecosystems';
-import { FailedToRunTestError } from '../src/lib/errors/failed-to-run-test-error';
+import { loadJson } from '../../utils';
+import * as featureFlags from '../../../src/lib/feature-flags';
+import { config as userConfig } from '../../../src/lib/user-config';
+import * as analysis from '../../../src/lib/plugins/sast/analysis';
+import { Options, TestOptions } from '../../../src/lib/types';
+import * as ecosystems from '../../../src/lib/ecosystems';
+import { FailedToRunTestError } from '../../../src/lib/errors/failed-to-run-test-error';
 const { getCodeAnalysisAndParseResults } = analysis;
 const osName = require('os-name');
 
@@ -21,17 +21,20 @@ describe('Test snyk code', () => {
   const failedCodeTestMessage = "Failed to run 'code test'";
   const fakeApiKey = '123456789';
   const sampleSarifResponse = loadJson(
-    __dirname + '/fixtures/sast/sample-sarif.json',
+    path.join(__dirname, '/../../fixtures/sast/sample-sarif.json'),
   );
   const sampleAnalyzeFoldersResponse = loadJson(
-    __dirname + '/fixtures/sast/sample-analyze-folders-response.json',
+    path.join(
+      __dirname,
+      '/../../fixtures/sast/sample-analyze-folders-response.json',
+    ),
   );
 
   const isWindows =
     osName()
       .toLowerCase()
       .indexOf('windows') === 0;
-  const fixturePath = path.join(__dirname, 'fixtures', 'sast');
+  const fixturePath = path.join(__dirname, '../../fixtures', 'sast');
   const cwd = process.cwd();
 
   function readFixture(filename: string) {
